[Update] プレイヤーのHPをfloat型に変更 ObjectManagerにGetPlayerHp関数作成

+++ DirectX12 Game Solution/DirectX12 MainProject/Classes/Collision/ObjectManager.cpp
@@ -30,7 +30,7 @@ BoundingOrientedBox ObjectManager::GetBossLHandCollision() {
 	return boss_->GetLHandCollision();
 }
 
-int ObjectManager::GetPlayerHP() {
+float ObjectManager::GetPlayerHP() {
 	return player_->GetPlayerHP();
 }
