+++ DirectX12 Game Solution/DirectX12 MainProject/Classes/Collision/ObjectManager.h
@@ -17,7 +17,7 @@ public:
 	BoundingOrientedBox GetBossCoreCollision();
 	BoundingOrientedBox GetBossRHandCollision();
 	BoundingOrientedBox GetBossLHandCollision();
-	int GetPlayerHP();
+	float GetPlayerHP();
 	bool GetPlayerAttackFlag();
 	bool GetBossRAttackFlag();
 	bool GetBossLAttackFlag();
